Refactoring __init__ for swarm
